[feat] Handle exceptions in server
Exceptions in server should be gracefully handled and notified to the
client, so that the client can handle the case. Basically, there are
two types of exception: 1) ServerError, and 2) ClientError.

ServerError is raised by an incorrect implementation in server, which
should be handled by server logic.
ClientError is raised by an incorrect input from the client.

Signed-off-by: JaewonHur <hurjaewon@snu.ac.kr>