[fix] Redefine exception types
There are two types of exceptions: 1) PrimeError, and 2) UserError.

The root cause of PrimeErro is on the prime implementation, and it
should be transparently handled by prime.
UserError results from an incorrect user input, and it should be
reported back to the user.

Prime handles and filters as many inputs as possible in the client
frontend.

Signed-off-by: JaewonHur <hurjaewon@snu.ac.kr>