[feat] Raise AttributeError in Prime server
We opt to hide any information other than the name (used for
referencing) of the variable in DE. It means even the type of the
variable is unknown and just assumed from FE, and forward exception from
the server on the AttributeError.
It is important in terms of security, since the type info can also be a
sort of information leak. At least Prime can track such leakge by
monitoring the exceptions in server.

Signed-off-by: JaewonHur <hurjaewon@snu.ac.kr>