[feat] Do now allow hash() on Proxy
Any computation on the tainted object should result in tainted output
(i.e., allocated in Prime server).
For the case of container types, sequences do not require computation on
the actual value, so they don't have to be in server. Sets and maps need
hash computation, so they should be allocated in server.
Prohibit hash computation on Proxy objects to remove any inconsistency.

Signed-off-by: JaewonHur <hurjaewon@snu.ac.kr>