[feat] Check trusted objects and methods
Prime may executes codes from 1) functions defined in trusted modules
(e.g., builtins, pytorch), 2) methods of a class instance from trusted
modules, and 3) functions or methods of an instance of a class, whose
prototypes are explicitly exported using ExportDef.

1) Prime checks the root of the functions before invoking it.
2) Prime checks the root of the type of the instance before allocating
a class instance. If the root is in trusted packages, it allows without
analyzing the prototype. (But still need to check overriden method or
other things...)
3) Prime analyzes the exported codes.

Signed-off-by: JaewonHur <hurjaewon@snu.ac.kr>